chatbot-service/requirements.txt
numpy<2.0

--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1,<2.3

sentence-transformers>=2.2
faiss-cpu>=1.7

# ---------- HuggingFace base ----------
transformers>=4.39

langchain>=0.1,<0.2
langchain-community>=0.0.30
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6

fastapi>=0.104
uvicorn[standard]>=0.24
python-multipart>=0.0.6
python-dotenv>=1.0
requests>=2.31
pydantic>=2.5,<3
pypdf>=3.17

# ---------- LLM SDK ----------
google-generativeai>=0.3
